+++ CHANGES.md
@@ -24,6 +24,7 @@
   - ➕ `padEnd` / `pad_end`
   - ➕ `padStart` / `pad_start`
   - ➕ `repeat`
+  - ➕ `replace` => `jsReplace` / `js_replace`
 
 ## 0.0.1 - 2023-04-28